+++ src/main/java/net/pincette/json/streams/TestProvider.java
@@ -9,4 +9,6 @@ interface TestProvider<T, U, V, W> extends Provider<T, U, V, W> {
   void createTopics(final Set<String> topics);
 
   void deleteTopics(final Set<String> topics);
+
+  Producer getProducer();
 }